chore: Clone V2 Engine to avoid touching V1 dependencies
